refactor: 로그인/회원가입/마이페이지 폴더 구조 및 파일명 수정
