+++ src/pages/my-page/components/EmptyData/index.js
