+++ src/pages/my-page/components/MyProfile/ProgressBar/index.js
